backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.6
face_recognition
dlib
rembg>=2.0.50
onnxruntime>=1.16.0
opencv-python-headless>=4.9.0
numpy>=1.26.0
pillow>=10.2.0
sqlalchemy>=2.0.25
pydantic>=2.5.3
setuptools
